Decimal edit

+++ HotelGuru/HotelGuru.DataContext/Dtos/InvoiceDto.cs
@@ -10,8 +10,8 @@ namespace HotelGuru.DataContext.Dtos
     public class InvoiceDto
     {
         public int Id { get; set; }
-        public int Price { get; set; }
-        public int ExtraPrice { get; set; }
-        public int TotalPrice => Price + ExtraPrice;
+        public decimal Price { get; set; }
+        public decimal ExtraPrice { get; set; }
+        public decimal TotalPrice => Price + ExtraPrice;
     }
 }
